requirements.txt
# Core training stack. Pinning versions avoids Gym/NumPy/JAX breakage.
numpy<2
gym==0.23.1
jax==0.6.2
jaxlib==0.6.2
jax-cuda12-pjrt==0.6.2 ; platform_system == "Linux"
jax-cuda12-plugin==0.6.2 ; platform_system == "Linux"
flax==0.10.7
optax==0.2.8
tqdm>=4.66,<5
ml_collections==1.1.0
wandb==0.26.0
dm-control==1.0.3.post1
dmcgym @ git+https://github.com/ikostrikov/dmcgym
pybullet==3.2.7
attrs==26.1.0
filterpy==1.4.5
pillow>=10,<13
tensorflow-cpu==2.21.0
tensorflow-probability==0.25.0
